Effects.Exception

Description

Exception handling. The interface here combines ideas from the following three libraries in a somewhat idiosyncratic way:

annotated-exception

exceptions

safe-exceptions

We have the following goals:

Typeclass abstraction: exceptions's MonadThrow/MonadCatch/MonadMask hierarchy.

Throw exceptions w/ CallStack: Inspired by annotated-exception, we use a custom type ExceptionCS for this purpose. This functionality may be removed once GHC natively supports combining exceptions and CallStack (tentatively GHC 9.8). See the following proposal for more information: https://github.com/ghc-proposals/ghc-proposals/pull/330

Throw/catch exceptions in accordance with safe-exceptions: That is, do not throw or catch async exceptions. We do not actually re-export safe-exceptions functions; the functions here are bespoke, primarily to add HasCallStack constraints.

Masking/bracket uses mask, not uninterruptibleMask: We take the position that uninterruptibleMask should not be the default behavior, as legitimate uses are rare, and for those corner cases, deadlocks are worse than interruptible operations receiving async exceptions.

The API exported here is not comprehensive vis-à-vis any of the mentioned libraries, so if more functionality is required, a direct dependency will be necessary.

Note that, like safe-exceptions, the typeclass methods for MonadThrow (throwM) and MonadCatch (catch) are not exported here because they are overridden to prevent catching async exceptions. This means that if one needs to manually write an instance for any of those classes, then a dependency on exceptions is required.

Since: 0.1

Global mechanisms

class Monad m => MonadGlobalException m where

Effect for global exception mechanisms.

Since: 0.1

Methods

setUncaughtExceptionHandler :: HasCallStack => (SomeException -> m ()) -> m ()

Lifted setUncaughtExceptionHandler.

Since: 0.1

getUncaughtExceptionHandler :: HasCallStack => m (SomeException -> m ())

Lifted getUncaughtExceptionHandler.

Since: 0.1

setUncaughtExceptionDisplay :: (HasCallStack, MonadGlobalException m) => (String -> m ()) -> m ()

setUncaughtExceptionHandler with displayCSNoMatchHandler i.e. calls displayException on any uncaught exceptions, passing the result to the param handler. ExitSuccess is ignored.

Since: 0.1

setUncaughtExceptionDisplayCSNoMatch :: (HasCallStack, MonadGlobalException m) => [ExceptionProxy] -> (String -> m ()) -> m ()

Like setUncaughtExceptionDisplay, except callstacks are not printed for exceptions matching the param proxies a la displayCSNoMatch.

Since: 0.1

CallStack

The callstack API and implementation is heavily inspired by the excellent annotated-exception library.

The primary difference is that annotated-exception is more general and comprehensive; the primary exception type allows one to tie arbitrary data to its exceptions.

By comparison, the functionality here is only concerned with CallStack. That enables both the implementation and usage to be simpler, if you only have the modest goal of obtaining exception callstacks.

Another difference is that ExceptionCS's Show instance uses its displayException, which pretty-prints the CallStack. This decision is due to many functions/ibraries defaulting to Show, so this way is more conservative if you really want stacktraces, at the cost of some transparency.

Note that if an exception e is thrown with throwCS, you can no longer catch it with normal catch @e, as it is now a ExceptionCS e. You can instead use catchCS @e.

data ExceptionCS e

Attaches a CallStack to an arbitrary exception. The Show instance uses displayException i.e. the underlying exceptions' displayException and pretty-prints the CallStack due to some libraries (e.g. testing) defaulting to Show when printing exceptions.

Since: 0.1

Constructors

MkExceptionCS e CallStack

throwCS :: forall m e a. (Exception e, HasCallStack, MonadThrow m) => e -> m a

Wraps an exception in ExceptionCS and rethrows. If the e is also an ExceptionCS, the callStacks are merged.

Since: 0.1

catchCS :: forall m e a. (Exception e, HasCallStack, MonadCatch m) => m a -> (e -> m a) -> m a

Catches both e and ExceptionCS e. The given handler is wrapped in addCS.

Since: 0.1

catchAnyCS :: forall m a. (HasCallStack, MonadCatch m) => m a -> (SomeException -> m a) -> m a

catchCS specialized to all synchronous exceptions.

Since: 0.1

handleCS :: forall m e a. (Exception e, HasCallStack, MonadCatch m) => (e -> m a) -> m a -> m a

Flipped catchCS.

Since: 0.1

handleAnyCS :: forall m a. (HasCallStack, MonadCatch m) => (SomeException -> m a) -> m a -> m a

handleCS specialized to SomeException.

Since: 0.1

tryCS :: forall m e a. (Exception e, MonadCatch m) => m a -> m (Either e a)

Try for catchCS.

Since: 0.1

tryAnyCS :: forall m a. MonadCatch m => m a -> m (Either SomeException a)

tryCS specialized to SomeException.

Since: 0.1

addCS :: forall m a. (HasCallStack, MonadCatch m) => m a -> m a

Turns any caught exceptions e into an ExceptionCS e with attached CallStack and rethrows.

Since: 0.1

addOuterCS :: forall m a. (HasCallStack, MonadCatch m) => CallStack -> m a -> m a

Like addCS, except it merges the given "outer callstack" with the one generated by a caught exception.

Since: 0.1

Utils

data ExceptionProxy

Proxy for exception types. Used with displayCSNoMatch.

Since: 0.1

Constructors

forall e.Exception e => MkExceptionProxy (Proxy e)

displayNoCS :: forall e. Exception e => e -> String

Like displayException, except it has specific logic to skip any found CallStacks.

Since: 0.1

displayCSNoMatch :: Exception e => [ExceptionProxy] -> e -> String

displayCSNoMatch proxies e is equivalent to displayNoCS if e matches the type of some proxy in p. Otherwise calls displayException.

This can be useful if we ordinarily do not want to print callstacks, doing so only in an unexpected case. For instance, we can use the following to ensure we do not print callstacks for specific Ex1 and Ex2:

data Ex1 = ... deriving anyclass Exception
data Ex2 = ... deriving anyclass Exception

let  displayEx' :: SomeException -> String
     displayEx' =
       displayCSNoMatch
         [ MkExceptionProxy (Proxy @Ex1),
           MkExceptionProxy (Proxy @Ex2)
         ]

-- in main
setUncaughtExceptionHandler displayEx'

Since: 0.1

displayCSNoMatchHandler :: Applicative f => [ExceptionProxy] -> (String -> f ()) -> SomeException -> f ()

Calls displayCSNoMatch on any uncaught exceptions, passing the result to the param handler. ExitSuccess is ignored.

Since: 0.1

Basic exceptions

The functionality here is a mix of exceptions and safe-exceptions. On the one hand, we reuse exceptions' typeclass hierarchy. On the other hand, the functions we export (e.g. throwM, catch) are reimplemented with safe-exceptions' philosophy: by default, do not throw or catch async exceptions. The only reason we do not directly export all safe-exceptions' functions is due to them missing HasCallStack constraints.

Throwing

class Monad m => MonadThrow (m :: Type -> Type)

A class for monads in which exceptions may be thrown.

Instances should obey the following law:

throwM e >> x = throwM e

In other words, throwing an exception short-circuits the rest of the monadic computation.

Minimal complete definition

throwM

throwM :: HasCallStack => (MonadThrow m, Exception e) => e -> m a

Like throwM but any thrown asynchronous exceptions will be thrown synchronously via toSyncException.

Since: 0.1

throwString :: (MonadThrow m, HasCallStack) => String -> m a

A convenience function for throwing a user error. This is useful for cases where it would be too high a burden to define your own exception type.

This throws an exception of type StringException. When GHC supports it (base 4.9 and GHC 8.0 and onward), it includes a call stack.

Since: safe-exceptions-0.1.5.0

Catching

class MonadThrow m => MonadCatch (m :: Type -> Type)

A class for monads which allow exceptions to be caught, in particular exceptions which were thrown by throwM.

Instances should obey the following law:

catch (throwM e) f = f e

Note that the ability to catch an exception does not guarantee that we can deal with all possible exit points from a computation. Some monads, such as continuation-based stacks, allow for more than just a success/failure strategy, and therefore catch cannot be used by those monads to properly implement a function such as finally. For more information, see MonadMask.

Minimal complete definition

catch

catch :: forall m e a. HasCallStack => (MonadCatch m, Exception e) => m a -> (e -> m a) -> m a

Like upstream catch, but will not catch asynchronous exceptions.

Since: 0.1

catchAny :: forall m a. HasCallStack => MonadCatch m => m a -> (SomeException -> m a) -> m a

catch specialized to all synchronous exceptions.

Since: 0.1

handle :: forall m e a. HasCallStack => (MonadCatch m, Exception e) => (e -> m a) -> m a -> m a

Flipped version of catch.

Since: 0.1

handleAny :: forall m a. HasCallStack => MonadCatch m => (SomeException -> m a) -> m a -> m a

Flipped version of catchAny.

Since: 0.1

try :: forall m e a. HasCallStack => (MonadCatch m, Exception e) => m a -> m (Either e a)

Like upstream try, but will not catch asynchronous exceptions.

Since: 0.1

tryAny :: forall m a. HasCallStack => MonadCatch m => m a -> m (Either SomeException a)

try specialized to catch all synchronous exceptions.

Since: 0.1

data Handler (m :: Type -> Type) a

Generalized version of Handler

Constructors

Exception e => Handler (e -> m a)

catches :: forall m a. HasCallStack => MonadCatch m => m a -> [Handler m a] -> m a

Like upstream catches, but will not catch asynchronous exceptions.

Since: 0.1

onException :: forall m a b. (HasCallStack, MonadCatch m) => m a -> m b -> m a

Run an action only if an exception is thrown in the main action. The exception is not caught, simply rethrown.

NOTE The action is only run if an exception is thrown. If the monad supports other ways of aborting the computation, the action won't run if those other kinds of errors are thrown. See onError.

NOTE This is not the onException re-exported from either exceptions or safe-exceptions. Why? The exceptions variant reuses its catch, which does not have the async exception behavior we want (i.e. do not catch async exceptions). The safe-exceptions version, on the other hand, use its bracket functions, meaning it performs uninterruptibleMask, which we also do not want.

Hence this version, which is based on catchAny i.e. it does not catch any asynchronous exceptions, nor does it invoke any masking.

Since: 0.1

catchIOError :: (HasCallStack, MonadCatch m) => m a -> (IOError -> m a) -> m a

Catch all IOError (eqv. IOException) exceptions. Still somewhat too general, but better than using catchAll. See catchIf for an easy way of catching specific IOErrors based on the predicates in System.IO.Error.

Masking

This section directly reexports exceptions' functions: we favor its mask over safe-exceptions' uninterruptibleMask.

class MonadCatch m => MonadMask (m :: Type -> Type) where

A class for monads which provide for the ability to account for all possible exit points from a computation, and to mask asynchronous exceptions. Continuation-based monads are invalid instances of this class.

Instances should ensure that, in the following code:

fg = f `finally` g

The action g is called regardless of what occurs within f, including async exceptions. Some monads allow f to abort the computation via other effects than throwing an exception. For simplicity, we will consider aborting and throwing an exception to be two forms of "throwing an error".

If f and g both throw an error, the error thrown by fg depends on which errors we're talking about. In a monad transformer stack, the deeper layers override the effects of the inner layers; for example, ExceptT e1 (Except e2) a represents a value of type Either e2 (Either e1 a), so throwing both an e1 and an e2 will result in Left e2. If f and g both throw an error from the same layer, instances should ensure that the error from g wins.

Effects other than throwing an error are also overridden by the deeper layers. For example, StateT s Maybe a represents a value of type s -> Maybe (a, s), so if an error thrown from f causes this function to return Nothing, any changes to the state which f also performed will be erased. As a result, g will see the state as it was before f. Once g completes, f's error will be rethrown, so g' state changes will be erased as well. This is the normal interaction between effects in a monad transformer stack.

By contrast, lifted-base's version of finally always discards all of g's non-IO effects, and g never sees any of f's non-IO effects, regardless of the layer ordering and regardless of whether f throws an error. This is not the result of interacting effects, but a consequence of MonadBaseControl's approach.

Methods

mask :: HasCallStack => ((forall a. m a -> m a) -> m b) -> m b

Runs an action with asynchronous exceptions disabled. The action is provided a method for restoring the async. environment to what it was at the mask call. See Control.Exception's mask.

uninterruptibleMask :: HasCallStack => ((forall a. m a -> m a) -> m b) -> m b

Like mask, but the masked computation is not interruptible (see Control.Exception's uninterruptibleMask. WARNING: Only use if you need to mask exceptions around an interruptible operation AND you can guarantee the interruptible operation will only block for a short period of time. Otherwise you render the program/thread unresponsive and/or unkillable.

generalBracket

Arguments

:: HasCallStack

=> m a
acquire some resource

-> (a -> ExitCase b -> m c)
release the resource, observing the outcome of the inner action

-> (a -> m b)
inner action to perform with the resource

-> m (b, c)

A generalized version of bracket which uses ExitCase to distinguish the different exit cases, and returns the values of both the use and release actions. In practice, this extra information is rarely needed, so it is often more convenient to use one of the simpler functions which are defined in terms of this one, such as bracket, finally, onError, and bracketOnError.

This function exists because in order to thread their effects through the execution of bracket, monad transformers need values to be threaded from use to release and from release to the output value.

NOTE This method was added in version 0.9.0 of this library. Previously, implementation of functions like bracket and finally in this module were based on the mask and uninterruptibleMask functions only, disallowing some classes of tranformers from having MonadMask instances (notably multi-exit-point transformers like ExceptT). If you are a library author, you'll now need to provide an implementation for this method. The StateT implementation demonstrates most of the subtleties:

generalBracket acquire release use = StateT $ s0 -> do
  ((b, _s2), (c, s3)) <- generalBracket
    (runStateT acquire s0)
    ((resource, s1) exitCase -> case exitCase of
      ExitCaseSuccess (b, s2) -> runStateT (release resource (ExitCaseSuccess b)) s2

      -- In the two other cases, the base monad overrides use's state
      -- changes and the state reverts to s1.
      ExitCaseException e     -> runStateT (release resource (ExitCaseException e)) s1
      ExitCaseAbort           -> runStateT (release resource ExitCaseAbort) s1
    )
    ((resource, s1) -> runStateT (use resource) s1)
  return ((b, c), s3)

The StateT s m implementation of generalBracket delegates to the m implementation of generalBracket. The acquire, use, and release arguments given to StateT's implementation produce actions of type StateT s m a, StateT s m b, and StateT s m c. In order to run those actions in the base monad, we need to call runStateT, from which we obtain actions of type m (a, s), m (b, s), and m (c, s). Since each action produces the next state, it is important to feed the state produced by the previous action to the next action.

In the ExitCaseSuccess case, the state starts at s0, flows through acquire to become s1, flows through use to become s2, and finally flows through release to become s3. In the other two cases, release does not receive the value s2, so its action cannot see the state changes performed by use. This is fine, because in those two cases, an error was thrown in the base monad, so as per the usual interaction between effects in a monad transformer stack, those state changes get reverted. So we start from s1 instead.

Finally, the m implementation of generalBracket returns the pairs (b, s) and (c, s). For monad transformers other than StateT, this will be some other type representing the effects and values performed and returned by the use and release actions. The effect part of the use result, in this case _s2, usually needs to be discarded, since those effects have already been incorporated in the release action.

The only effect which is intentionally not incorporated in the release action is the effect of throwing an error. In that case, the error must be re-thrown. One subtlety which is easy to miss is that in the case in which use and release both throw an error, the error from release should take priority. Here is an implementation for ExceptT which demonstrates how to do this.

generalBracket acquire release use = ExceptT $ do
  (eb, ec) <- generalBracket
    (runExceptT acquire)
    (eresource exitCase -> case eresource of
      Left e -> return (Left e) -- nothing to release, acquire didn't succeed
      Right resource -> case exitCase of
        ExitCaseSuccess (Right b) -> runExceptT (release resource (ExitCaseSuccess b))
        ExitCaseException e       -> runExceptT (release resource (ExitCaseException e))
        _                         -> runExceptT (release resource ExitCaseAbort))
    (either (return . Left) (runExceptT . use))
  return $ do
    -- The order in which we perform those two Either effects determines
    -- which error will win if they are both Lefts. We want the error from
    -- release to win.
    c <- ec
    b <- eb
    return (b, c)

Since: exceptions-0.9.0

data ExitCase a

A MonadMask computation may either succeed with a value, abort with an exception, or abort for some other reason. For example, in ExceptT e IO you can use throwM to abort with an exception (ExitCaseException) or throwE to abort with a value of type e (ExitCaseAbort).

Constructors

ExitCaseSuccess a

ExitCaseException SomeException

ExitCaseAbort

mask_ :: (HasCallStack, MonadMask m) => m a -> m a

Like mask, but does not pass a restore action to the argument.

uninterruptibleMask_ :: (HasCallStack, MonadMask m) => m a -> m a

Like uninterruptibleMask, but does not pass a restore action to the argument.

bracket :: (HasCallStack, MonadMask m) => m a -> (a -> m c) -> (a -> m b) -> m b

Generalized abstracted pattern of safe resource acquisition and release in the face of errors. The first action "acquires" some value, which is "released" by the second action at the end. The third action "uses" the value and its result is the result of the bracket.

If an error is thrown during the use, the release still happens before the error is rethrown.

Note that this is essentially a type-specialized version of generalBracket. This function has a more common signature (matching the signature from Control.Exception), and is often more convenient to use. By contrast, generalBracket is more expressive, allowing us to implement other functions like bracketOnError.

bracket_ :: (HasCallStack, MonadMask m) => m a -> m c -> m b -> m b

Version of bracket without any value being passed to the second and third actions.

finally :: (HasCallStack, MonadMask m) => m a -> m b -> m a

Perform an action with a finalizer action that is run, even if an error occurs.

bracketOnError :: (HasCallStack, MonadMask m) => m a -> (a -> m c) -> (a -> m b) -> m b

Like bracket, but only performs the final action if an error is thrown by the in-between computation.

Utils

data SyncExceptionWrapper

Wrap up an asynchronous exception to be treated as a synchronous exception

This is intended to be created via toSyncException

Since: safe-exceptions-0.1.0.0

data AsyncExceptionWrapper

Wrap up a synchronous exception to be treated as an asynchronous exception

This is intended to be created via toAsyncException

Since: safe-exceptions-0.1.0.0

toSyncException :: Exception e => e -> SomeException

Convert an exception into a synchronous exception

For synchronous exceptions, this is the same as toException. For asynchronous exceptions, this will wrap up the exception with SyncExceptionWrapper

Since: safe-exceptions-0.1.0.0

toAsyncException :: Exception e => e -> SomeException

Convert an exception into an asynchronous exception

For asynchronous exceptions, this is the same as toException. For synchronous exceptions, this will wrap up the exception with AsyncExceptionWrapper

Since: safe-exceptions-0.1.0.0

isSyncException :: Exception e => e -> Bool

Check if the given exception is synchronous

Since: safe-exceptions-0.1.0.0

isAsyncException :: Exception e => e -> Bool

Check if the given exception is asynchronous

Since: safe-exceptions-0.1.0.0

Exiting

These functions represent Exit. die can be found in Effects.FileSystem.HandleWriter.

data ExitCode

Defines the exit codes that a program can return.

Constructors

ExitSuccess
indicates successful termination;

ExitFailure Int
indicates program failure with an exit code. The exact interpretation of the code is operating-system dependent. In particular, some values may be prohibited (e.g. 0 on a POSIX-compliant system).

Functions

exitFailure :: HasCallStack => MonadThrow m => m a

The computation exitFailure is equivalent to exitWith (ExitFailure exitfail), where exitfail is implementation-dependent.

Since: 0.1

exitSuccess :: HasCallStack => MonadThrow m => m a

The computation exitSuccess is equivalent to exitWith ExitSuccess, It terminates the program successfully.

Since: 0.1

exitWith :: HasCallStack => MonadThrow m => ExitCode -> m a

Computation exitWith code throws ExitCode code. Normally this terminates the program, returning code to the program's caller.

On program termination, the standard Handles stdout and stderr are flushed automatically; any other buffered Handles need to be flushed manually, otherwise the buffered data will be discarded.

A program that fails in any other way is treated as if it had called exitFailure. A program that terminates successfully without calling exitWith explicitly is treated as if it had called exitWith ExitSuccess.

As an ExitCode is not an IOException, exitWith bypasses the error handling in the IO monad and cannot be intercepted by catch from the Prelude. However it is a SomeException, and can be caught using the functions of Control.Exception. This means that cleanup computations added with bracket (from Control.Exception) are also executed properly on exitWith.

Note: in GHC, exitWith should be called from the main program thread in order to exit the process. When called from another thread, exitWith will throw an ExitException as normal, but the exception will not cause the process itself to exit.

Since: 0.1

Misc

data CallStack

CallStacks are a lightweight method of obtaining a partial call-stack at any point in the program.

A function can request its call-site with the HasCallStack constraint. For example, we can define

putStrLnWithCallStack :: HasCallStack => String -> IO ()

as a variant of putStrLn that will get its call-site and print it, along with the string given as argument. We can access the call-stack inside putStrLnWithCallStack with callStack.

>>> :{
putStrLnWithCallStack :: HasCallStack => String -> IO ()
putStrLnWithCallStack msg = do
  putStrLn msg
  putStrLn (prettyCallStack callStack)
:}

Thus, if we call putStrLnWithCallStack we will get a formatted call-stack alongside our string.

>>> putStrLnWithCallStack "hello"
hello
CallStack (from HasCallStack):
  putStrLnWithCallStack, called at <interactive>:... in interactive:Ghci...

GHC solves HasCallStack constraints in three steps:

If there is a CallStack in scope -- i.e. the enclosing function has a HasCallStack constraint -- GHC will append the new call-site to the existing CallStack.

If there is no CallStack in scope -- e.g. in the GHCi session above -- and the enclosing definition does not have an explicit type signature, GHC will infer a HasCallStack constraint for the enclosing definition (subject to the monomorphism restriction).

If there is no CallStack in scope and the enclosing definition has an explicit type signature, GHC will solve the HasCallStack constraint for the singleton CallStack containing just the current call-site.

CallStacks do not interact with the RTS and do not require compilation with -prof. On the other hand, as they are built up explicitly via the HasCallStack constraints, they will generally not contain as much information as the simulated call-stacks maintained by the RTS.

A CallStack is a [(String, SrcLoc)]. The String is the name of function that was called, the SrcLoc is the call-site. The list is ordered with the most recently called function at the head.

NOTE: The intrepid user may notice that HasCallStack is just an alias for an implicit parameter ?callStack :: CallStack. This is an implementation detail and should not be considered part of the CallStack API, we may decide to change the implementation in the future.

Since: base-4.8.1.0

class (Typeable e, Show e) => Exception e where

Any type that you wish to throw or catch as an exception must be an instance of the Exception class. The simplest case is a new exception type directly below the root:

data MyException = ThisException | ThatException
    deriving Show

instance Exception MyException

The default method definitions in the Exception class do what we need in this case. You can now throw and catch ThisException and ThatException as exceptions:

*Main> throw ThisException `catch` \e -> putStrLn ("Caught " ++ show (e :: MyException))
Caught ThisException

In more complicated examples, you may wish to define a whole hierarchy of exceptions:

---------------------------------------------------------------------
-- Make the root exception type for all the exceptions in a compiler

data SomeCompilerException = forall e . Exception e => SomeCompilerException e

instance Show SomeCompilerException where
    show (SomeCompilerException e) = show e

instance Exception SomeCompilerException

compilerExceptionToException :: Exception e => e -> SomeException
compilerExceptionToException = toException . SomeCompilerException

compilerExceptionFromException :: Exception e => SomeException -> Maybe e
compilerExceptionFromException x = do
    SomeCompilerException a <- fromException x
    cast a

---------------------------------------------------------------------
-- Make a subhierarchy for exceptions in the frontend of the compiler

data SomeFrontendException = forall e . Exception e => SomeFrontendException e

instance Show SomeFrontendException where
    show (SomeFrontendException e) = show e

instance Exception SomeFrontendException where
    toException = compilerExceptionToException
    fromException = compilerExceptionFromException

frontendExceptionToException :: Exception e => e -> SomeException
frontendExceptionToException = toException . SomeFrontendException

frontendExceptionFromException :: Exception e => SomeException -> Maybe e
frontendExceptionFromException x = do
    SomeFrontendException a <- fromException x
    cast a

---------------------------------------------------------------------
-- Make an exception type for a particular frontend compiler exception

data MismatchedParentheses = MismatchedParentheses
    deriving Show

instance Exception MismatchedParentheses where
    toException   = frontendExceptionToException
    fromException = frontendExceptionFromException

We can now catch a MismatchedParentheses exception as MismatchedParentheses, SomeFrontendException or SomeCompilerException, but not other types, e.g. IOException:

*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: MismatchedParentheses))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: SomeFrontendException))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: SomeCompilerException))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: IOException))
*** Exception: MismatchedParentheses

Minimal complete definition

Nothing

Methods

toException :: e -> SomeException

fromException :: SomeException -> Maybe e

displayException :: e -> String

Render this exception value in a human-friendly manner.

Default implementation: show.

Since: base-4.8.0.0

type HasCallStack = ?callStack :: CallStack

Request a CallStack.

NOTE: The implicit parameter ?callStack :: CallStack is an implementation detail and should not be considered part of the CallStack API, we may decide to change the implementation in the future.

Since: base-4.9.0.0

data SomeException

The SomeException type is the root of the exception type hierarchy. When an exception of type e is thrown, behind the scenes it is encapsulated in a SomeException.

data IOException

Exceptions that occur in the IO monad. An IOException records a more specific error type, a descriptive string and maybe the handle that was used when the error was flagged.

data StringException

Exception type thrown by throwString.

Note that the second field of the data constructor depends on GHC/base version. For base 4.9 and GHC 8.0 and later, the second field is a call stack. Previous versions of GHC and base do not support call stacks, and the field is simply unit (provided to make pattern matching across GHC versions easier).

Since: safe-exceptions-0.1.5.0