Effects.Exception.Annotation.CallStack.Legacy

Description

Deprecated: The legacy CallStack exception annotations are deprecated asthe functionality is subsumed in base 4.20 (GHC 9.10).

Exception handling. The interface here combines ideas from the following three libraries in a somewhat idiosyncratic way:

annotated-exception

exceptions

safe-exceptions

We have the following goals:

Typeclass abstraction: exceptions's MonadThrow/MonadCatch/MonadMask hierarchy.

Throw exceptions w/ CallStack: Inspired by annotated-exception, we use a custom type ExceptionCS for this purpose. This functionality may be removed once GHC natively supports combining exceptions and CallStack (tentatively GHC 9.8). See the following proposal for more information: https://github.com/ghc-proposals/ghc-proposals/pull/330

Throw/catch exceptions in accordance with safe-exceptions: That is, do not throw or catch async exceptions. We do not actually re-export safe-exceptions functions; the functions here are bespoke, primarily to add HasCallStack constraints.

Masking/bracket uses mask, not uninterruptibleMask: We take the position that uninterruptibleMask should not be the default behavior, as legitimate uses are rare, and for those corner cases, deadlocks are worse than interruptible operations receiving async exceptions.

The API exported here is not comprehensive vis-à-vis any of the mentioned libraries, so if more functionality is required, a direct dependency will be necessary.

Note that, like safe-exceptions, the typeclass methods for MonadThrow (throwM) and MonadCatch (catch) are not exported here because they are overridden to prevent catching async exceptions. This means that if one needs to manually write an instance for any of those classes, then a dependency on exceptions is required.

Since: 0.1

Global mechanisms

setUncaughtExceptionDisplay :: (HasCallStack, MonadGlobalException m) => (String -> m ()) -> m ()

setUncaughtExceptionHandler with displayCSNoMatchHandler i.e. calls displayException on any uncaught exceptions, passing the result to the param handler. ExitSuccess is ignored.

Since: 0.1

setUncaughtExceptionDisplayCSNoMatch :: (HasCallStack, MonadGlobalException m) => [ExceptionProxy] -> (String -> m ()) -> m ()

Like setUncaughtExceptionDisplay, except callstacks are not printed for exceptions matching the param proxies a la displayCSNoMatch.

Since: 0.1

CallStack

The callstack API and implementation is heavily inspired by the excellent annotated-exception library.

The primary difference is that annotated-exception is more general and comprehensive; the primary exception type allows one to tie arbitrary data to its exceptions.

By comparison, the functionality here is only concerned with CallStack. That enables both the implementation and usage to be simpler, if you only have the modest goal of obtaining exception callstacks.

Another difference is that ExceptionCS's Show instance uses its displayException, which pretty-prints the CallStack. This decision is due to many functions/ibraries defaulting to Show, so this way is more conservative if you really want stacktraces, at the cost of some transparency.

Note that if an exception e is thrown with throwCS, you can no longer catch it with normal catch @e, as it is now a ExceptionCS e. You can instead use catchCS @e.

data ExceptionCS e

Attaches a CallStack to an arbitrary exception. The Show instance uses displayException i.e. the underlying exceptions' displayException and pretty-prints the CallStack due to some libraries (e.g. testing) defaulting to Show when printing exceptions.

Since: 0.1

Constructors

MkExceptionCS e CallStack

throwCS :: forall m e a. (Exception e, HasCallStack, MonadThrow m) => e -> m a

Wraps an exception in ExceptionCS and rethrows. If the e is also an ExceptionCS, the callStacks are merged.

Since: 0.1

catchCS :: forall m e a. (Exception e, HasCallStack, MonadCatch m) => m a -> (e -> m a) -> m a

Catches both e and ExceptionCS e. The given handler is wrapped in addCS.

Since: 0.1

catchAnyCS :: (HasCallStack, MonadCatch m) => m a -> (SomeException -> m a) -> m a

catchCS specialized to all synchronous exceptions.

Since: 0.1

handleCS :: forall m e a. (Exception e, HasCallStack, MonadCatch m) => (e -> m a) -> m a -> m a

Flipped catchCS.

Since: 0.1

handleAnyCS :: (HasCallStack, MonadCatch m) => (SomeException -> m a) -> m a -> m a

handleCS specialized to SomeException.

Since: 0.1

tryCS :: forall m e a. (Exception e, MonadCatch m) => m a -> m (Either e a)

Try for catchCS.

Since: 0.1

tryAnyCS :: MonadCatch m => m a -> m (Either SomeException a)

tryCS specialized to SomeException.

Since: 0.1

addCS :: (HasCallStack, MonadCatch m) => m a -> m a

Turns any caught exceptions e into an ExceptionCS e with attached CallStack and rethrows.

Since: 0.1

addOuterCS :: (HasCallStack, MonadCatch m) => CallStack -> m a -> m a

Like addCS, except it merges the given "outer callstack" with the one generated by a caught exception.

Since: 0.1

Utils

data ExceptionProxy

Proxy for exception types. Used for matching multiple exception types

Since: 0.1

Constructors

Exception e => MkExceptionProxy (Proxy e)

displayNoCS :: Exception e => e -> String

Like displayException, except it has specific logic to skip any found CallStacks.

Since: 0.1

displayCSNoMatch :: Exception e => [ExceptionProxy] -> e -> String

displayCSNoMatch proxies e is equivalent to displayNoCS if e matches the type of some proxy in p. Otherwise calls displayException.

This can be useful if we ordinarily do not want to print callstacks, doing so only in an unexpected case. For instance, we can use the following to ensure we do not print callstacks for specific Ex1 and Ex2:

data Ex1 = ... deriving anyclass Exception
data Ex2 = ... deriving anyclass Exception

let  displayEx' :: SomeException -> String
     displayEx' =
       displayCSNoMatch
         [ MkExceptionProxy (Proxy @Ex1),
           MkExceptionProxy (Proxy @Ex2)
         ]

-- in main
setUncaughtExceptionHandler displayEx'

Since: 0.1

displayCSNoMatchHandler :: Applicative f => [ExceptionProxy] -> (String -> f ()) -> SomeException -> f ()

Calls displayCSNoMatch on any uncaught exceptions, passing the result to the param handler. ExitSuccess is ignored.

Since: 0.1